Recover Interface::unlock() function distinct from cleanup().
Nate points out it's hard to write a cleanup() function if you're
not sure whether the data structure is locked or not!

It would *probably* be better/faster to combine cleanup() with
unlock() (no unlock, lock, cleanup, unlock). But this is easier,
and there were bugs with the prior combination, so do this for now.