Always use the transaction hash table.
Even for small transaction sets.

The key innovation here is not to clear the transaction hashtable
between transactions. We do this by assuming that every transaction
contains <= 32768 items. We keep track of a moving boundary, the
"hash base". If the value of a hash table element is <= the "hash
base", then that element is empty. The "hash base" is updated after
every transaction to account for all the previous transaction's
items. We clear the hash table memory only when the "hash base"
reaches 32768 (so we might overflow).

This is good for 5% on ./concurrent 3 hash.