Recover "chunked" design that means TItems occupy stable memory.
Even if we go from 512->513 items.

I disabled this earlier for a couple reasons. We had gone with the
indexed approach (i.e. TransProxy == Transaction + index), but that
added overhead every time a TransProxy was used. I really wanted that
TItem pointer for vector predicates!!

So go back to this style, in which pointers stay stable. Like the
indexed approach, there was a small performance hit---<1% but nonzero.
But make performance better (on ./concurrent 3 array -j1) by passing
the writeset into stop() in case of commit. With that hack (which, sad
to say, we could do for the previous design too), we are a little more
than 1% faster on that benchmark.