Adding skeleton to new performance experiment
The new experiment, called 'concurrent_blue-red-warehouse_experiment',
will first populate the database with random keys and a number of blue
or red items, where the number of blue items is calculated from input
'percent_blue', and where the total number of items is equal to input
'num_items'. Then the database will be partitioned into a number of
equal pieces equal to input 'num_warehouses'. Then, a number of threads
specified by 'num_threads' will each be run with a transaction that
decides on a 'warehouse' to work on, and then executes random operations
on that 'warehouse'.

Right now the experiment correctly parses the items.