Modify skeleton for changed experiment
The new experiment will still be concurrent, but it will not work with blue-red
items. Since the experiment is meant to work somewhat like the TPCC benchmark,
each item will contain several variables. Inputs for changing the size of the
database were removed since the size is meant to be fixed at 10,000 items. Pblue
input was removed because we are no longer working with blue-red items.