Fix Invalidation of Items, Minor Fixes
The main fix was reimplementing the invalidation of TransItems when
aborting a subaction of a two phase transaction. The reason is because
after a transaction reaches a certain tset size in Sto, a new chunk to
store more TransItems gets allocated. The past implementation depended
on all TransItems being contiguous in memory, and this commit removes
that assumption. Formatting and removal of dead code is included in this
commit.