Correct error in Wikipedia updatePage transaction.
Previously, the transaction ran as follows:

   START
   INSERT NEW TEXT
   INSERT NEW REVISION
   UPDATE PAGE TABLE
   INSERT RECENT CHANGES
   SELECT WATCHING USERS
   If there are any watching users, then:
      COMMIT
      START
      INSERT WATCHLIST ROWS FOR ALL WATCHING USERS
      COMMIT
      START
   INSERT LOG
   UPDATE USER
   COMMIT

Couple weird things about this. Why is INSERT LOG, UPDATE USER
only in the main transaction if there are no watching users?
But the serious problem with CU was that the WATCHLIST ROWS CUs
might never commit if a concurrent deleteWatch transaction deleted
a corresponding user. Because of our shiny new write validation
for CUs, that delete prevented the transaction from committing,
but the watchlist was never reloaded, so txns spun forever.